Добавлен каркас экрана "Список мест"
